Refactoring src: done some refactoring on src for cli pourpose
